+++ code/feature_extraction/extract_sketch_token_features.py
@@ -52,7 +52,7 @@ def extract_features(subject, use_node_storage=False, debug=False, which_prf_gri
     
     t = time.time()
     with h5py.File(fn2save, 'w') as data_set:
-        dset = data_set.create_dataset("features", np.shape(features_each_prf), dtype=np.float64)
+        dset = data_set.create_dataset("features", np.shape(features_each_prf), dtype=np.float32)
         data_set['/features'][:,:,:] = features_each_prf
         data_set.close()  
     elapsed = time.time() - t
